+++ modules/material/material.service.js
@@ -10,7 +10,7 @@ const {
 const {
   HISTORY_ACTIONS: { ADD_EVENT, DELETE_EVENT, EDIT_EVENT },
   HISTORY_NAMES: { MATERIAL_EVENT },
-} = require('../../consts/history-actions');
+} = require('../../consts/history-events');
 const createTranslations = require('../../utils/createTranslations');
 const {
   addTranslations,
